sheep: avoid diskfull caused by recovery process
sheep can corrupt its cluster by diskfull with recovery process. For
avoiding this problem, this patch adds a new option -F to sheep. If
this command is passed to the sheep process, every sheep process of
the cluster skips recovery if there is a possibility of diskfull
during recovery.

Fixes https://github.com/sheepdog/sheepdog/issues/59

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>